+++ test/uk/gov/hmrc/tai/model/domain/EmploymentSpec.scala
@@ -16,15 +16,185 @@
 
 package uk.gov.hmrc.tai.model.domain
 
-import java.time.LocalDate
 import org.scalatestplus.play.PlaySpec
+import play.api.libs.json.Json
 import uk.gov.hmrc.tai.model.domain.income.Live
 import uk.gov.hmrc.tai.model.tai.TaxYear
 
+import java.time.LocalDate
+
 class EmploymentSpec extends PlaySpec {
+  private val taxYear = TaxYear()
 
-  "Employment" must {
+  private def createAnnualAccount(rtiStatus: RealTimeStatus, taxYear: TaxYear = taxYear): AnnualAccount =
+    AnnualAccount(0, taxYear, rtiStatus, Nil, Nil)
 
+  private val singleEmploymentWithAllRefs = List(
+    Employment(
+      "XXX PPPP",
+      Live,
+      Some("64765"),
+      LocalDate.of(2016, 4, 6),
+      None,
+      Nil,
+      "754",
+      "AZ00070",
+      2,
+      Some(100),
+      false,
+      false
+    )
+  )
+
+  // TODO: 9180 Why are below no longer used? Check in main
+//  private val singleEmploymentWithMissingPayrollNumber = List(
+//    Employment(
+//      "XXX PPPP",
+//      Live,
+//      None,
+//      LocalDate.of(2016, 4, 6),
+//      None,
+//      Nil,
+//      "754",
+//      "AZ00070",
+//      2,
+//      Some(100),
+//      false,
+//      false
+//    )
+//  )
+//  private val singleEmploymentWithEmptyStringPayrollNumber = List(
+//    Employment(
+//      "XXX PPPP",
+//      Live,
+//      Some(""),
+//      LocalDate.of(2016, 4, 6),
+//      None,
+//      Nil,
+//      "754",
+//      "AZ00070",
+//      2,
+//      Some(100),
+//      false,
+//      false
+//    )
+//  )
+//  private val dualEmployment = List(
+//    Employment(
+//      "XXX PPPP",
+//      Live,
+//      Some("64765"),
+//      LocalDate.of(2016, 4, 6),
+//      None,
+//      Nil,
+//      "754",
+//      "AZ00070",
+//      2,
+//      Some(100),
+//      false,
+//      false
+//    ),
+//    Employment(
+//      "XXX PPPP",
+//      Live,
+//      Some("64766"),
+//      LocalDate.of(2016, 4, 6),
+//      None,
+//      Nil,
+//      "754",
+//      "AZ00070",
+//      2,
+//      Some(100),
+//      false,
+//      false
+//    )
+//  )
+
+  "Format" must {
+    "un-marshall and marshal valid Employment json and object" when {
+      "given a valid Employment object and json" in {
+        val employmentDetails = List(
+          Employment(
+            "TEST",
+            Live,
+            Some("12345"),
+            LocalDate.parse("2017-05-26"),
+            None,
+            List(
+              AnnualAccount(
+                0,
+                TaxYear(2017),
+                Available,
+                List(Payment(LocalDate.parse("2017-05-26"), 10, 10, 10, 10, 10, 10, Monthly, Some(true))),
+                List(
+                  EndOfTaxYearUpdate(LocalDate.parse("2017-05-26"), List(Adjustment(NationalInsuranceAdjustment, 10)))
+                )
+              )
+            ),
+            "",
+            "",
+            2,
+            Some(100),
+            false,
+            false
+          )
+        )
+
+        val json = Json.arr(
+          Json.obj(
+            "name"             -> "TEST",
+            "employmentStatus" -> Live.toString,
+            "payrollNumber"    -> "12345",
+            "startDate"        -> "2017-05-26",
+            "annualAccounts" -> Json.arr(
+              Json.obj(
+                "sequenceNumber" -> 0,
+                "taxYear"        -> 2017,
+                "realTimeStatus" -> "Available",
+                "payments" -> Json.arr(
+                  Json.obj(
+                    "date"                              -> "2017-05-26",
+                    "amountYearToDate"                  -> 10,
+                    "taxAmountYearToDate"               -> 10,
+                    "nationalInsuranceAmountYearToDate" -> 10,
+                    "amount"                            -> 10,
+                    "taxAmount"                         -> 10,
+                    "nationalInsuranceAmount"           -> 10,
+                    "payFrequency"                      -> "Monthly",
+                    "duplicate"                         -> true
+                  )
+                ),
+                "endOfTaxYearUpdates" -> Json.arr(
+                  Json.obj(
+                    "date" -> "2017-05-26",
+                    "adjustments" -> Json.arr(
+                      Json.obj(
+                        "type"   -> "NationalInsuranceAdjustment",
+                        "amount" -> 10
+                      )
+                    )
+                  )
+                )
+              )
+            ),
+            "taxDistrictNumber"            -> "",
+            "payeNumber"                   -> "",
+            "sequenceNumber"               -> 2,
+            "cessationPay"                 -> 100,
+            "hasPayrolledBenefit"          -> false,
+            "receivingOccupationalPension" -> false
+          )
+        )
+
+        Json.toJson(employmentDetails) mustBe json
+
+        json.as[Seq[Employment]] mustBe employmentDetails
+
+      }
+    }
+  }
+
+  "tempUnavailableStubExistsForYear" must {
     "return true if any annual account exists, with a status of temporarily unavailable, for the requested year" in {
       val accounts =
         Seq(createAnnualAccount(TemporarilyUnavailable), createAnnualAccount(Available, taxYear))
@@ -42,7 +212,9 @@ class EmploymentSpec extends PlaySpec {
 
       employmentWithAccount.tempUnavailableStubExistsForYear(taxYear) mustBe false
     }
+  }
 
+  "hasAnnualAccountsForYear" must {
     "return true if an annual accounts exist for a particular year" in {
       val accounts = Seq(createAnnualAccount(Available, TaxYear(2019)), createAnnualAccount(Available))
       val employmentWithAccount = singleEmploymentWithAllRefs.head.copy(annualAccounts = accounts)
@@ -56,7 +228,9 @@ class EmploymentSpec extends PlaySpec {
 
       employmentWithAccount.hasAnnualAccountsForYear(taxYear) mustBe false
     }
+  }
 
+  "annualAccountsForYear" must {
     "return all annual accounts for the request year" in {
       val expectedAccount1 = createAnnualAccount(Available)
       val expectedAccount2 = createAnnualAccount(Unavailable)
@@ -78,89 +252,5 @@ class EmploymentSpec extends PlaySpec {
     }
 
   }
-
-  val taxYear = TaxYear()
-
-  def createAnnualAccount(rtiStatus: RealTimeStatus, taxYear: TaxYear = taxYear): AnnualAccount =
-    AnnualAccount(0, taxYear, rtiStatus, Nil, Nil)
-
-  val singleEmploymentWithAllRefs = List(
-    Employment(
-      "XXX PPPP",
-      Live,
-      Some("64765"),
-      LocalDate.of(2016, 4, 6),
-      None,
-      Nil,
-      "754",
-      "AZ00070",
-      2,
-      Some(100),
-      false,
-      false
-    )
-  )
-  val singleEmploymentWithMissingPayrollNumber = List(
-    Employment(
-      "XXX PPPP",
-      Live,
-      None,
-      LocalDate.of(2016, 4, 6),
-      None,
-      Nil,
-      "754",
-      "AZ00070",
-      2,
-      Some(100),
-      false,
-      false
-    )
-  )
-  val singleEmploymentWithEmptyStringPayrollNumber = List(
-    Employment(
-      "XXX PPPP",
-      Live,
-      Some(""),
-      LocalDate.of(2016, 4, 6),
-      None,
-      Nil,
-      "754",
-      "AZ00070",
-      2,
-      Some(100),
-      false,
-      false
-    )
-  )
-  val dualEmployment = List(
-    Employment(
-      "XXX PPPP",
-      Live,
-      Some("64765"),
-      LocalDate.of(2016, 4, 6),
-      None,
-      Nil,
-      "754",
-      "AZ00070",
-      2,
-      Some(100),
-      false,
-      false
-    ),
-    Employment(
-      "XXX PPPP",
-      Live,
-      Some("64766"),
-      LocalDate.of(2016, 4, 6),
-      None,
-      Nil,
-      "754",
-      "AZ00070",
-      2,
-      Some(100),
-      false,
-      false
-    )
-  )
 
 }
